+++ packages/components/src/FilterBar/subcomponents/index.ts
@@ -1,0 +1,4 @@
+export * from "./FilterBarDatePicker"
+export * from "./FilterBarDateRangePicker"
+export * from "./FilterBarMultiSelect"
+export * from "./FilterBarSelect"