+++ packages/components/src/FilterMultiSelect/subcomponents/SectionDivider/SectionDivider.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+import { Divider } from "@kaizen/draft-divider"
+import styles from "./SectionDivider.module.scss"
+
+export const SectionDivider = (): JSX.Element => (
+  <Divider classNameOverride={styles.divider} variant="canvas" />
+)
+
+SectionDivider.displayName = "FilterMultiSelect.SectionDivider"