+++ packages/components/src/Workflow/subcomponents/Header/components/Actions/index.tsx
@@ -1,2 +1,0 @@
-export * from "./Actions"
-export * from "./WorkflowExit"