+++ packages/rich-text-editor/docs/RichTextContent.stories.tsx
@@ -5,7 +5,7 @@ import dummyContent from "./dummyContent.json"
 
 export default {
   tags: ["autodocs"],
-  title: "Components/Rich Text Content",
+  title: "Components/Rich Text Editor/Rich Text Content",
   component: RichTextContent,
   parameters: {
     docs: {